@@ -13,11 +13,12 @@
   "dependencies": {
     "chalk": "^5.4.1",
     "dotenv": "^16.4.7",
+    "ejs": "^3.1.10",
     "express": "^5.0.1",
     "nodemon": "^3.1.9",
     "zod": "^3.24.1"
   },
   "devDependencies": {
-    "eslint": "^9.17.0"
+    "eslint": "^9.18.0"
   }
 }
