@@ -1,6 +1,6 @@
-//*====================================================
+//*================================================
 //* npm Commands :- 
-//*===================================================
+//*================================================
 //! npm install <package-name>@<version> 
 //? Example :- 
     //* npm install express@4.0.0 // ^4.0.0
