@@ -9,7 +9,7 @@ import readline from "readline";
 const rl = readline.createInterface({
     input : process.stdin, 
     output : process.stdout
-})
+});
 
 
 const todos = [];
